pyproject.toml
[project]
name = "flight_monitor"
version = "0.1.0"
description = "Flight data monitoring service"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flightradarapi = {version = "^1.3.34", extras = ["requests"]}
psycopg2-binary = "^2.9.10"
sqlalchemy = {version = "^2.0.40", extras = ["asyncio"]}
alembic = "^1.15.2"
geopy = "^2.4.1"
pytz = "^2025.2"
matplotlib = "^3.10.1"
folium = "^0.19.5"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.8.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
ipython = "^8.25.0"

[tool.poetry.scripts]
collect-flights = "app.services.flightradar_services:main"
